+++ slang.h
@@ -1142,6 +1142,7 @@ namespace slang
 #include "source/core/slang-string.cpp"
 #include "source/core/stream.cpp"
 #include "source/core/text-io.cpp"
+#include "source/slang/ast-legalize.cpp"
 #include "source/slang/bytecode.cpp"
 #include "source/slang/diagnostics.cpp"
 #include "source/slang/dxc-support.cpp"
@@ -1157,7 +1158,6 @@ namespace slang
 #include "source/slang/preprocessor.cpp"
 #include "source/slang/profile.cpp"
 #include "source/slang/lookup.cpp"
-#include "source/slang/lower.cpp"
 #include "source/slang/lower-to-ir.cpp"
 #include "source/slang/check.cpp"
 #include "source/slang/compiler.cpp"
